finance-system/web/requirements.txt
Flask==3.0.0
duckdb==0.9.2
PyYAML==6.0.1
python-dotenv==1.0.0
